requirements.txt
python-binance==1.0.19
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
matplotlib>=3.5.0